permutation.cc,permutation.h,permutation_types.x: Added the UnbalancedLRPermutation class, and tested. Works well.
common/function-objs.h: Added a struct to print integer values in binary.

card/io_flat.h: Elaborate in a comment that the scalar versions of read()
and write() need to be called with exactly an index_t index, not int or
such.

card/io.cc: moved tests to separate file card/test-io.cc.